refactor: atualiza função de inserir filme
