+++ include/Estoque.hpp
@@ -4,17 +4,17 @@
 #include <iostream>
 #include <string.h>
 #include <vector>
-#include "./Filme.hpp"
+#include "Filme.hpp"
 #include <string.h>
 
 class Estoque
 {
 private:
-    std::vector<Filme> estoque;
+    std::vector<Filme*> estoque;
     std::string diretorio;
 public:
     void lerArquivo(std::string diretorio);
-    void inserirFilme(Filme filme);
+    void inserirFilme(Filme* filme);
     void removerFilme(int codigo);
     std::vector<Filme> pesquisarFilmes(std::string filtro);
     void imprimirRelatiorio();
